fix: satisfy lint line cap and ES2020 lib typecheck
Split the site-location index (locateSites, its per-paragraph/per-row
memos and the tracked-row anchors) out of review-reads.ts into
review-site-locations.ts along the "where is this node" seam, bringing
the queue module back under the 1000-line cap. Declare WeakRef locally
in recent-root-cache.ts: it is ES2021 and the workspace tsconfigs pin
lib at ES2020, while every supported runtime ships it.
